\name{Catch-package}
\alias{Catch-package}
\alias{Catch}
\docType{package}
\title{
  The Catch C++ Automated Test Cases for C++ (and, maybe, C).
}
\description{
  Catch stands for C++ Automated Test Cases in Headers and is a
  multi-paradigm automated test framework for C++ and Objective-C (and,
  maybe, C).
}
\details{
\tabular{ll}{
Package: \tab Catch\cr
Type: \tab Package\cr
License: \tab BSL-1.0\cr
}
  A R package for providing Catch header file.
}
\references{
https://github.com/philsquared/Catch
}
\keyword{ package }
\examples{
# None
}
