\name{Catch-package}
\alias{Catch-package}
\alias{Catch}
\docType{package}
\title{
  The Catch C++ Automated Test Cases for C++ (and, maybe, C).
}
\description{
  Catch stands for C++ Automated Test Cases in Headers and is a
  multi-paradigm automated test framework for C++ and Objective-C (and,
  maybe, C).
}
\details{
\tabular{ll}{
Package: \tab Catch\cr
Type: \tab Package\cr
License: \tab BSL-1.0\cr
}
  Boost provides free peer-reviewed portable C++ source
  libraries with a large part implemented as template headers.
  We place a number of these libraries in this package as a more efficient
  distribution system for CRAN.  The idea is that you can write
  a package that depends on some subset of the Boost Libraries,
  yet you do not need to distribute a copy of this code with your
  package.
}
\references{
https://github.com/philsquared/Catch
}
\keyword{ package }
\examples{
# None
}
