\name{Catch-package}
\alias{Catch-package}
\alias{Catch}
\docType{package}
\title{
  The Catch C++ Automated Test Cases for C++ (and, maybe, C).
}
\description{
  Catch stands for C++ Automated Test Cases in Headers and is a
  multi-paradigm automated test framework for C++ and Objective-C (and,
  maybe, C).
}
\details{
\tabular{ll}{
Package: \tab Catch\cr
Type: \tab Package\cr
License: \tab BSL-1.0\cr
}
A R package for providing Catch header file. The macro WARN is renamed
to CWARN, unless CATCH_NO_RENAME_WARN_TO_CWARN is defined, since WANR is
also defined in 'include/R_ext/RS.h'.
}
\references{
https://github.com/philsquared/Catch
}
\keyword{ package }
\examples{
# None
}
